+++ app/Http/Controllers/FluxoController.php
@@ -3,11 +3,16 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use App\Fluxo;
 
 class FluxoController extends Controller {
 
     public function index() {
+
+        if (!Auth::check()) {
+            return redirect('/');
+        }
 
         //$areas = Area::where('user_id', auth()->user()->id)->get();
         $fluxos = Fluxo::paginate(10);
@@ -17,6 +22,10 @@ class FluxoController extends Controller {
 
     public function create() {
 
+        if (!Auth::check()) {
+            return redirect('/');
+        }
+
         // indica inclusão
         $acao = 1;
 
@@ -24,6 +33,10 @@ class FluxoController extends Controller {
     }
 
     public function store(Request $request) {
+
+        if (!Auth::check()) {
+            return redirect('/');
+        }
 
         // recupera todos os campos do formulário
         $dados = $request->all();
@@ -41,6 +54,10 @@ class FluxoController extends Controller {
 
     public function edit($id) {
 
+        if (!Auth::check()) {
+            return redirect('/');
+        }
+
         // obtém os dados do registro a ser editado 
         $reg = Fluxo::find($id);
 
@@ -56,6 +73,10 @@ class FluxoController extends Controller {
     }
 
     public function update(Request $request, $id) {
+
+        if (!Auth::check()) {
+            return redirect('/');
+        }
 
         $reg = Fluxo::find($id);
 
@@ -74,6 +95,11 @@ class FluxoController extends Controller {
     }
 
     public function destroy($id) {
+
+        if (!Auth::check()) {
+            return redirect('/');
+        }
+
         $fluxo = Fluxo::find($id);
         if ($fluxo->delete()) {
             return redirect()->route('fluxos.index')
@@ -83,11 +109,20 @@ class FluxoController extends Controller {
 
     public function pesq() {
 
+        if (!Auth::check()) {
+            return redirect('/');
+        }
+
         $carros = Carro::paginate(3);
         return view('carros_pesq', compact('carros'));
     }
 
     public function filtro(Request $request) {
+
+        if (!Auth::check()) {
+            return redirect('/');
+        }
+
         // obtém dados do form de pesquisa
         $modelo = $request->modelo;
         $precomax = $request->precomax;
@@ -104,6 +139,11 @@ class FluxoController extends Controller {
     }
 
     public function filtro2(Request $request) {
+
+        if (!Auth::check()) {
+            return redirect('/');
+        }
+
         // obtém dados do form de pesquisa
         $modelo = $request->modelo;
         $precomax = $request->precomax;
@@ -119,6 +159,11 @@ class FluxoController extends Controller {
     }
 
     public function graf() {
+
+        if (!Auth::check()) {
+            return redirect('/');
+        }
+
         $carros = DB::table('carros')
                 ->join('marcas', 'carros.marca_id', '=', 'marcas.id')
                 ->select('marcas.nome as marca', DB::raw('count(*) as num'))
